+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/procs/compare/v0.10.7...Unreleased) - ReleaseDate
 
+* [Changed] Add `LESSCHARSET=utf-8` by default [#75](https://github.com/dalance/procs/issues/75)
+
 ## [v0.10.7](https://github.com/dalance/procs/compare/v0.10.6...v0.10.7) - 2020-11-05
 
 * [Added] detect_width config [#76](https://github.com/dalance/procs/issues/76)